Refactor functions into proper domains
