+++ lib/wallaby/dsl.ex
@@ -3,12 +3,8 @@ defmodule Wallaby.DSL do
 
   defmacro __using__([]) do
     quote do
-      import Wallaby.DSL.Navigation
-      import Wallaby.DSL.Finders
-      import Wallaby.DSL.Attributes
-      import Wallaby.DSL.Actions
-      import Wallaby.DSL.Matchers
-      import Wallaby.DSL.Helpers
+      import Wallaby.Node
+      import Wallaby.Session
     end
   end
 end
