+++ test/support/server_case.ex
@@ -1,13 +1,13 @@
 defmodule Wallaby.ServerCase do
   use ExUnit.CaseTemplate
 
-  setup _tags do
+  setup_all _tags do
     {:ok, server} = Wallaby.TestServer.start
 
     on_exit fn ->
       Wallaby.TestServer.stop(server)
     end
-    
+
     {:ok, %{server: server}}
   end
 end
