+++ test/wallaby/xpath_test.exs
@@ -1,5 +1,3 @@
 defmodule Wallaby.XPathTest do
   use ExUnit.Case
-
-  import Wallaby.XPath
 end
